+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 10/02/2020
+
+* Updated MSAL.js to 1.4.1 (resolves id token cache issue).
+* Added unsubscribe to broadcast events on ngOnDestroy().
+
 ## 08/31/2020
 
 * Updated MSAL.js to 1.4.0 and MSAL-Angular to 1.1.1.